SECOND RELEASE ★ Better built-in file browser ★ Better peer list ★ Optional download folder ★ Search function ★ Language support: Hungarian, German & Russian
Bug fixes:
X There was no reaction when connection settings changed
X Torrent was opening again when orientation changed

+++ trunk/DrTorrent/DrTorrent/src/hu/bute/daai/amorg/drtorrent/coding/bencode/BencodedInteger.java
@@ -8,7 +8,7 @@ public class BencodedInteger extends Bencoded {
 	private long value_;
 
 	/** Creates a new instance of MTBencodedInteger */
-	public BencodedInteger(int value) {
+	public BencodedInteger(long value) {
 		value_ = value;
 	}
